requirements.txt
django
django-oscar
django-oscar-api
sorl-thumbnail
django-treebeard
django-haystack
django-tables2
Pillow
djangorestframework
djangorestframework-simplejwt
